refactor: move generator function
