+++ cmd/lotto-numbers/main.go
@@ -13,9 +13,6 @@ import (
 func main() {
 	log.Printf("Version %q", version.Version)
 
-	num := generateLotteryNumbers([]int{1, 1, 1, 1, 1, 3, 5, 7, 9, 11, 13, 15, 17, 19, 21, 23, 25, 27, 29}, 5, 5)
-	log.Printf("Numbers: %v", num)
-
 	// Set up the HTTP server using Gorilla Mux
 	r := mux.NewRouter()
 	r.HandleFunc("/lottery-numbers", handlers.GetLotteryNumbers).Methods("GET")
